+++ all-is-cubes/src/space/tests.rs
@@ -6,7 +6,7 @@ use cgmath::EuclideanSpace as _;
 use indoc::indoc;
 
 use crate::block::{
-    Block, BlockDef, BlockDefTransaction, EvalBlockError, Primitive, Resolution::*, AIR,
+    Atom, Block, BlockDef, BlockDefTransaction, EvalBlockError, Primitive, Resolution::*, AIR,
 };
 use crate::content::make_some_blocks;
 use crate::drawing::VoxelBrush;
@@ -477,7 +477,7 @@ fn set_physics_light_rays() {
 #[test]
 fn block_tick_action() {
     let [mut block1, block2] = make_some_blocks();
-    if let Primitive::Atom(attributes, _) = block1.primitive_mut() {
+    if let Primitive::Atom(Atom { attributes, .. }) = block1.primitive_mut() {
         attributes.tick_action = Some(VoxelBrush::single(block2.clone()));
     } else {
         panic!();